Added exceptions and handlers for api.

+++ opinions_app/api_views.py
@@ -1,7 +1,7 @@
 from flask import jsonify, request
-from random import randrange
 
 from . import app, db
+from .error_handlers import InvalidAPIUsage
 from .models import Opinion
 from .views import random_opinion
 
@@ -10,7 +10,9 @@ from .views import random_opinion
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
     # Получаем объект по id или выбрасываем ошибку 404:
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage(('Мнение с указанным id не найдено', 404))
     # Конвертируем данные в JSON и возвращаем JSON-объект и HTTP-код ответа:
     return jsonify({'opinion': opinion.to_dict()}), 200
 
@@ -18,9 +20,16 @@ def get_opinion(id):
 @app.route('/api/opinions/<int:id>/', methods=['PATCH'])
 def update_opinion(id):
     data = request.get_json()
+    if (
+        'text' in data and
+        Opinion.query.filter_by(text=data['text']).first() is not None
+    ):
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     # Если метод get_or_404 не найдёт указанный ID, 
     # то он выбросит исключение 404:
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     opinion.title = data.get('title', opinion.title)
     opinion.text = data.get('text', opinion.text)
     opinion.source = data.get('source', opinion.source)
@@ -29,12 +38,14 @@ def update_opinion(id):
     # Объект opinion добавлять в сессию не нужно.
     # Этот объект получен из БД методом get_or_404() и уже хранится в сессии.
     db.session.commit()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    return jsonify({'opinion': opinion.to_dict()}), 201
 
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     return '', 204
@@ -50,7 +61,16 @@ def get_opinions():
 @app.route('/api/opinions/', methods=['POST'])
 def add_opinion():
     # Получение данных из запроса в виде словаря:
-    data = request.get_json()
+    data = request.get_json(silent=True)
+    if data is None:
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
+    if 'title' not in data and 'text' not in data:
+        # return jsonify({'error': 'В запросе отсутствуют обязательные поля'}), 400
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля')
+    # Если в базе данных уже есть объект 
+    # с таким же значением поля text...
+    if Opinion.query.filter_by(text=data['text']).first() is not None:
+        raise InvalidAPIUsage('Такое мнение уже есть в базе данных')
     # Создание нового пустого экземпляра модели:
     opinion = Opinion()
     # Наполнение экземпляра данными из запроса:
@@ -63,4 +83,7 @@ def add_opinion():
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    # Тут код ответа нужно указать явным образом.
+    raise InvalidAPIUsage('В базе данных нет мнений', 404)
